news_collector/news_collector/requirements.txt
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
beautifulsoup4==4.6.3
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.4.2
cssselect==1.0.3
feedfinder2==0.0.4
feedparser==5.2.1
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
jieba3k==0.35.1
lxml==4.2.5
mysqlclient==1.3.14
newspaper3k==0.2.8
nltk==3.4
parsel==1.5.1
Pillow==5.3.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
python-dateutil==2.7.5
PyYAML==3.13
queuelib==1.5.0
requests==2.20.1
requests-file==1.4.3
Scrapy==1.5.1
service-identity==17.0.0
singledispatch==3.4.0.3
six==1.11.0
tinysegmenter==0.3
tldextract==2.2.0
Twisted==18.9.0
urllib3==1.24.1
w3lib==1.19.0
zope.interface==4.6.0
